feat(security): paths_identical FS-identity helper + split fs tests (WI-3.1)
Add paths_identical(first, second) comparing OS file identity (Unix
dev+ino, Windows volume-serial+file-index) not path strings, for scoping
decisions that need "same file". Audited §3: drive scoping already uses
the typed DriveLetter and path_contains is a deliberate substring filter,
so no string-compare safety bug existed.

The cumulative Phase-A + WI-3.1 additions pushed fs.rs over the 800-LOC
policy ceiling, so the test module moved to fs/tests.rs via #[path]
(repo-idiomatic, mirrors runtime_dir/tests.rs). fs.rs now 625 LOC.

Tests: hardlink/symlink-target identical; distinct files not identical.
